# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenClarion API
  description: |
    Natural-language-driven alert governance platform.
    This specification is the single source of truth for all API contracts.
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/config/report-workflow-schedules:
    get:
      operationId: listReportWorkflowSchedules
      summary: List report workflow schedules
      description: Returns operator-managed report workflow schedule metadata. This read does not inspect Temporal Schedule state.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ListLimit"
      responses:
        "200":
          description: Report workflow schedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowScheduleListResponse"
              examples:
                schedules:
                  value:
                    items:
                      - id: 1
                        name: Daily report window
                        report_workflow_policy_id: 1
                        temporal_schedule_id: openclarion-report-policy-1-daily
                        interval_seconds: 86400
                        offset_seconds: 21600
                        replay_window_seconds: 3600
                        replay_delay_seconds: 300
                        replay_limit: 10000
                        catchup_window_seconds: 3600
                        enabled: false
                        enabled_at: null
                        disabled_at: null
                        created_at: "2026-06-06T02:00:00Z"
                        updated_at: "2026-06-06T02:00:00Z"
        "400":
          description: Report workflow schedule list parameters are invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-limit:
                  value:
                    error: "limit must be between 1 and 500"
        "500":
          description: Report workflow schedule list failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: list report workflow schedules failed
    post:
      operationId: createReportWorkflowSchedule
      summary: Create a report workflow schedule
      description: Stores report workflow schedule metadata as a disabled draft. This operation does not register a Temporal Schedule, start workflows, call alert providers, resolve secrets, or send notifications.
      tags:
        - config
      requestBody:
        description: Report workflow schedule metadata.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportWorkflowScheduleWriteRequest"
            examples:
              daily:
                value:
                  name: Daily report window
                  report_workflow_policy_id: 1
                  temporal_schedule_id: openclarion-report-policy-1-daily
                  interval_seconds: 86400
                  offset_seconds: 21600
                  replay_window_seconds: 3600
                  replay_delay_seconds: 300
                  replay_limit: 10000
                  catchup_window_seconds: 3600
      responses:
        "201":
          description: Report workflow schedule created as a disabled draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowSchedule"
              examples:
                created:
                  value:
                    id: 1
                    name: Daily report window
                    report_workflow_policy_id: 1
                    temporal_schedule_id: openclarion-report-policy-1-daily
                    interval_seconds: 86400
                    offset_seconds: 21600
                    replay_window_seconds: 3600
                    replay_delay_seconds: 300
                    replay_limit: 10000
                    catchup_window_seconds: 3600
                    enabled: false
                    enabled_at: null
                    disabled_at: null
                    created_at: "2026-06-06T02:00:00Z"
                    updated_at: "2026-06-06T02:00:00Z"
        "400":
          description: Report workflow schedule request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: "report workflow schedule: offset must be less than interval"
        "404":
          description: Bound report workflow policy was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing-binding:
                  value:
                    error: report workflow schedule binding not found
        "409":
          description: Report workflow schedule conflicts with an existing schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicate:
                  value:
                    error: report workflow schedule already exists
        "500":
          description: Report workflow schedule creation failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: create report workflow schedule failed
  /api/v1/config/report-workflow-schedules/{schedule_id}:
    get:
      operationId: getReportWorkflowSchedule
      summary: Get a report workflow schedule
      description: Returns one operator-managed report workflow schedule without querying Temporal.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ReportWorkflowScheduleID"
      responses:
        "200":
          description: Report workflow schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowSchedule"
              examples:
                schedule:
                  value:
                    id: 1
                    name: Daily report window
                    report_workflow_policy_id: 1
                    temporal_schedule_id: openclarion-report-policy-1-daily
                    interval_seconds: 86400
                    offset_seconds: 21600
                    replay_window_seconds: 3600
                    replay_delay_seconds: 300
                    replay_limit: 10000
                    catchup_window_seconds: 3600
                    enabled: true
                    enabled_at: "2026-06-06T02:05:00Z"
                    disabled_at: null
                    created_at: "2026-06-06T02:00:00Z"
                    updated_at: "2026-06-06T02:05:00Z"
        "404":
          description: Report workflow schedule was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: report workflow schedule not found
        "500":
          description: Report workflow schedule lookup failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: get report workflow schedule failed
    put:
      operationId: replaceReportWorkflowSchedule
      summary: Replace a report workflow schedule
      description: Replaces report workflow schedule metadata while preserving explicit enabled state. This operation does not register a Temporal Schedule, start workflows, call alert providers, resolve secrets, or send notifications.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ReportWorkflowScheduleID"
      requestBody:
        description: Replacement report workflow schedule metadata.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportWorkflowScheduleWriteRequest"
            examples:
              six-hour:
                value:
                  name: Six-hour report window
                  report_workflow_policy_id: 1
                  temporal_schedule_id: openclarion-report-policy-1-6h
                  interval_seconds: 21600
                  offset_seconds: 0
                  replay_window_seconds: 3600
                  replay_delay_seconds: 300
                  replay_limit: 10000
                  catchup_window_seconds: 1800
      responses:
        "200":
          description: Report workflow schedule replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowSchedule"
              examples:
                replaced:
                  value:
                    id: 1
                    name: Six-hour report window
                    report_workflow_policy_id: 1
                    temporal_schedule_id: openclarion-report-policy-1-6h
                    interval_seconds: 21600
                    offset_seconds: 0
                    replay_window_seconds: 3600
                    replay_delay_seconds: 300
                    replay_limit: 10000
                    catchup_window_seconds: 1800
                    enabled: false
                    enabled_at: null
                    disabled_at: null
                    created_at: "2026-06-06T02:00:00Z"
                    updated_at: "2026-06-06T02:10:00Z"
        "400":
          description: Report workflow schedule replacement request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: "report workflow schedule: replay_limit must be positive"
        "404":
          description: Replacement target or bound policy was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: report workflow schedule not found
        "409":
          description: Replacement schedule conflicts with an existing Temporal Schedule ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                duplicate:
                  value:
                    error: report workflow schedule already exists
        "500":
          description: Report workflow schedule replacement failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: replace report workflow schedule failed
  /api/v1/config/report-workflow-schedules/{schedule_id}/enable:
    post:
      operationId: enableReportWorkflowSchedule
      summary: Enable a report workflow schedule
      description: Explicitly enables a report workflow schedule after validating that the bound report workflow policy is enabled. This action does not register or unpause a Temporal Schedule.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ReportWorkflowScheduleID"
      responses:
        "200":
          description: Report workflow schedule enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowSchedule"
              examples:
                enabled:
                  value:
                    id: 1
                    name: Daily report window
                    report_workflow_policy_id: 1
                    temporal_schedule_id: openclarion-report-policy-1-daily
                    interval_seconds: 86400
                    offset_seconds: 21600
                    replay_window_seconds: 3600
                    replay_delay_seconds: 300
                    replay_limit: 10000
                    catchup_window_seconds: 3600
                    enabled: true
                    enabled_at: "2026-06-06T02:05:00Z"
                    disabled_at: null
                    created_at: "2026-06-06T02:00:00Z"
                    updated_at: "2026-06-06T02:05:00Z"
        "400":
          description: Report workflow schedule cannot be enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                disabled-policy:
                  value:
                    error: "report workflow schedule: report workflow policy must be enabled before schedule enablement"
        "404":
          description: Report workflow schedule or bound policy was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: report workflow schedule not found
        "500":
          description: Report workflow schedule enablement failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: enable report workflow schedule failed
  /api/v1/config/report-workflow-schedules/{schedule_id}/disable:
    post:
      operationId: disableReportWorkflowSchedule
      summary: Disable a report workflow schedule
      description: Explicitly disables a report workflow schedule. This action pauses future starts only after the later Temporal registration slice wires schedule state to Temporal; it does not cancel already-started report workflows.
      tags:
        - config
      parameters:
        - $ref: "#/components/parameters/ReportWorkflowScheduleID"
      responses:
        "200":
          description: Report workflow schedule disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportWorkflowSchedule"
              examples:
                disabled:
                  value:
                    id: 1
                    name: Daily report window
                    report_workflow_policy_id: 1
                    temporal_schedule_id: openclarion-report-policy-1-daily
                    interval_seconds: 86400
                    offset_seconds: 21600
                    replay_window_seconds: 3600
                    replay_delay_seconds: 300
                    replay_limit: 10000
                    catchup_window_seconds: 3600
                    enabled: false
                    enabled_at: null
                    disabled_at: "2026-06-06T02:15:00Z"
                    created_at: "2026-06-06T02:00:00Z"
                    updated_at: "2026-06-06T02:15:00Z"
        "400":
          description: Report workflow schedule disable request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid:
                  value:
                    error: schedule_id must be positive
        "404":
          description: Report workflow schedule was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: report workflow schedule not found
        "500":
          description: Report workflow schedule disablement failed server-side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server-error:
                  value:
                    error: disable report workflow schedule failed
...
components:
  parameters:
    ListLimit:
      name: limit
      in: query
      description: Maximum number of rows to return.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 500
        default: 100
...
    ReportWorkflowScheduleID:
      name: schedule_id
      in: path
      description: Report workflow schedule identifier.
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
...
    ReportWorkflowScheduleWriteRequest:
      type: object
      description: Report workflow schedule metadata accepted by create and replace operations. Enabled state is action-only.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          examples:
            - Daily report window
        report_workflow_policy_id:
          type: integer
          format: int64
          minimum: 1
          examples:
            - 1
        temporal_schedule_id:
          type: string
          minLength: 1
          maxLength: 200
          description: Submitted Temporal Schedule identifier; must not contain whitespace or control characters.
          examples:
            - openclarion-report-policy-1-daily
        interval_seconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 31536000
        offset_seconds:
          type: integer
          format: int64
          minimum: 0
          maximum: 31536000
        replay_window_seconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 31536000
        replay_delay_seconds:
          type: integer
          format: int64
          minimum: 0
          maximum: 31536000
        replay_limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 100000
        catchup_window_seconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 31536000
      required:
        - name
        - report_workflow_policy_id
        - temporal_schedule_id
        - interval_seconds
        - offset_seconds
        - replay_window_seconds
        - replay_delay_seconds
        - replay_limit
        - catchup_window_seconds
      additionalProperties: false
    ReportWorkflowScheduleListResponse:
      type: object
      description: Report workflow schedule list response.
      properties:
        items:
          type: array
          examples:
            -
              - id: 1
                name: Daily report window
                report_workflow_policy_id: 1
                temporal_schedule_id: openclarion-report-policy-1-daily
                interval_seconds: 86400
                offset_seconds: 21600
                replay_window_seconds: 3600
                replay_delay_seconds: 300
                replay_limit: 10000
                catchup_window_seconds: 3600
                enabled: false
                enabled_at: null
                disabled_at: null
                created_at: "2026-06-06T02:00:00Z"
                updated_at: "2026-06-06T02:00:00Z"
          items:
            $ref: "#/components/schemas/ReportWorkflowSchedule"
      required:
        - items
...
    ErrorResponse:
      type: object
      description: Error response payload.
      properties:
        error:
          type: string
          examples:
            - "limit must be between 1 and 500 (got 0)"
      required:
        - error
      examples:
        - error: "limit must be between 1 and 500 (got 0)"